LeaderboardSync/requirements.txt
gspread==3.7.0
oauth2client==4.1.3
pymongo==3.11.4
Flask==2.0.1
google-cloud-secret-manager==2.7.1
dnspython==2.1.0
